initial refactor complete.
