--- CarRental.Bookings/Calculator/Calculations/ICalculation.cs
@@ -1,12 +1,0 @@
-﻿using CarRental.Bookings.Entities;
-using CarRental.Vehicles;
-
-namespace CarRental.Bookings.Calculator.Calculations
-{
-    public interface ICalculation
-    {
-        float Run(IVehicle vehicle, int duration, float discount, Booking booking);
-
-        bool AppliesTo(IVehicle vehicle);
-    }
-}
