--- CarRental.Bookings/Exceptions/CannotBookCarException.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using CarRental.Bookings.Entities;
-
-namespace CarRental.Bookings.Exceptions
-{
-    public class CannotBookCarException : InvalidOperationException
-    {
-        public CannotBookCarException(string message) : base(message)
-        {
-        }
-
-        public CannotBookCarException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        public static CannotBookCarException New(Booking booking)
-        {
-            return new CannotBookCarException($"Vehicle {booking.CarId} is unavailable from {booking.RentalDate.Date} to {booking.ReturnDate.Date}");
-        }
-    }
-}
