--- CarRental.Bookings/Factory/IBookingFactory.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using CarRental.Bookings.Entities;
-using CarRental.Vehicles;
-
-namespace CarRental.Bookings.Factory
-{
-    public interface IBookingFactory
-    {
-        Booking NewBooking(IVehicle vehicle, DateTime start, int duration, string name);
-    }
-}
