+++ CarRental.Bookings/IBookingRepository.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace CarRental.Bookings
+{
+    public interface IBookingRepository
+    {
+        void AddBooking(Booking b);
+
+        List<Booking> GetBookings();
+    }
+}
