+++ CarRental.Bookings/Pricing/ICostCalculator.cs
@@ -1,0 +1,9 @@
+﻿using CarRental.Vehicles;
+
+namespace CarRental.Bookings.Pricing
+{
+    public interface ICostCalculator
+    {
+        float Calculate(IVehicle vehicle, int duration, float discount, Booking booking);
+    }
+}
