--- CarRental.Bookings/Validation/INewBookingValidator.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using CarRental.Bookings.Entities;
-
-namespace CarRental.Bookings.Validation
-{
-    public interface INewBookingValidator
-    {
-        bool CanBook(Booking booking, List<Booking> currentBookings);
-    }
-}
