--- CarRental.Bookings/Validation/IVehicleMaintenanceProvider.cs
@@ -1,7 +1,0 @@
-namespace CarRental.Bookings.Validation
-{
-    public interface IVehicleMaintenanceProvider
-    {
-        int DaysRequired(int vehicleId);
-    }
-}
