+++ CarRental.Bookings/lib/Factory/IBookingFactory.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using CarRental.Vehicles;
+
+namespace CarRental.Bookings.lib.Factory
+{
+    public interface IBookingFactory
+    {
+        Booking NewBooking(IVehicle vehicle, DateTime start, int duration, string name);
+    }
+}
