+++ CarRental.Bookings/lib/Validation/INewBookingValidator.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace CarRental.Bookings.lib.Validation
+{
+    public interface INewBookingValidator
+    {
+        bool CanBook(Booking booking, List<Booking> currentBookings);
+    }
+}
