+++ CarRental.Vehicles/lib/IVehicleMaintenanceProvider.cs
@@ -1,0 +1,7 @@
+namespace CarRental.Vehicles.lib
+{
+    public interface IVehicleMaintenanceProvider
+    {
+        int DaysRequired(int vehicleId);
+    }
+}
